add: bubble + restructure
