--- src/assets/Components/IncrementAnimation.module.css
@@ -1,51 +1,0 @@
-.bubble {
-  position: fixed;
-  top: 40%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  animation: floatUp 3s ease-out forwards;
-}
-
-.bubbleContent {
-  display: inline-block;
-  width: 100px;
-  height: 100px;
-  border-radius: 50%;
-  background-color: rgba(234, 151, 27, 0.5);
-  animation: bubbleGrow 3s ease-out forwards;
-  overflow: hidden;
-}
-
-.bubbleImage {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-}
-
-@keyframes floatUp {
-  from {
-    opacity: 1;
-    transform: translate(-50%, -50%) translateY(0);
-  }
-  to {
-    opacity: 0;
-    transform: translate(-50%, -50%) translateY(-300px);
-  }
-}
-
-@keyframes bubbleGrow {
-  0% {
-    transform: scale(0);
-  }
-  50% {
-    transform: scale(0.5);
-  }
-  80% {
-    transform: scale(0.5);
-    background-color: rgba(234, 151, 27, 1);
-  }
-  100% {
-    transform: scale(0);
-    background-color: rgba(234, 151, 27, 0);
-  }
-}
