+++ src/components/Background.jsx
@@ -1,5 +1,5 @@
-import sample from "./2.gif";
-import style from "./Background.module.css";
+import sample from "../assets/girl_standinf.gif";
+import style from "../styles/Background.module.css";
 function Background() {
   return (
     <div className={style.background_video}>
@@ -7,7 +7,7 @@ function Background() {
         <source src={sample} type="video/mp4" />
         Your browser does not support the video tag.
       </video> */}
-       <img src={sample} alt="Background GIF" className={style.gif} />
+      <img src={sample} alt="Background GIF" className={style.gif} />
     </div>
   );
 }
