+++ src/components/Counter.jsx
@@ -1,0 +1,3 @@
+function Counter({ count }) {
+  return <h3>Counter: {count} </h3>;
+}
